+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Json API pagination to `Illuminate\Database\Query\Builder` class
+
 ## [3.0.1] - 2021-01-20
 
 ## Added